+++ api/src/main/java/net/jqwik/api/RandomGenerator.java
@@ -32,7 +32,9 @@ public interface RandomGenerator<T> {
 
 		public abstract <T> RandomGenerator<T> unique(RandomGenerator<T> self);
 
-		public abstract <T> RandomGenerator<List<T>> collect(RandomGenerator<T> tRandomGenerator, Predicate<List<T>> until);
+		public abstract <T> RandomGenerator<List<T>> collect(RandomGenerator<T> self, Predicate<List<T>> until);
+
+		public abstract <T> RandomGenerator<T> injectDuplicates(RandomGenerator<T> self, double duplicateProbability);
 	}
 
 	/**
@@ -89,8 +91,14 @@ public interface RandomGenerator<T> {
 		return Stream.generate(() -> this.next(random));
 	}
 
+	@API(status = EXPERIMENTAL, since = "1.1.4")
 	default RandomGenerator<List<T>> collect(Predicate<List<T>> until) {
 		return RandomGeneratorFacade.implementation.collect(this, until);
+	}
+
+	@API(status = EXPERIMENTAL, since = "1.2.3")
+	default RandomGenerator<T> injectDuplicates(double duplicateProbability) {
+		return RandomGeneratorFacade.implementation.injectDuplicates(this, duplicateProbability);
 	}